requirements.txt
asgiref==3.11.1
Django==3.2.25
legacy-cgi==2.6.4
mysqlclient==2.2.8
python-dotenv==1.2.1
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2025.3
